requirements.txt
asgiref==3.3.1
Django==3.1.3
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-memcached==0.1.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
python-memcached==1.59
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
Markdown==3.2.2
django-filter==2.4.0
djangorestframework==3.11.0

